chore: rename to billboards instead of display ad
