+++ app/javascript/packs/billboardAfterRenderActions.js
